+++ tests/reader_test_plugins/init_error.py
@@ -1,0 +1,2 @@
+def init_reader(reader):
+    raise ValueError('someerror')